+++ src/jsMain/kotlin/com/foundryvtt/core/GameTime.kt
@@ -4,9 +4,10 @@ import io.socket.Socket
 import kotlin.js.Promise
 
 external class GameTime(socket: Socket) {
-    val serverTime: Int
-    val worldTime: Int
-    fun advance(seconds: Int, options: AnyObject = definedExternally): Promise<Int>
+    val serverTime: Double
+    val worldTime: Double
+    fun advance(seconds: Double, options: AnyObject = definedExternally): Promise<Double>
+    fun advance(seconds: Int, options: AnyObject = definedExternally): Promise<Double>
     fun sync(socket: Socket): Promise<GameTime>
-    fun onUpdateWorldTime(worldTime: Int, options: AnyObject, userId: String)
+    fun onUpdateWorldTime(worldTime: Double, options: AnyObject, userId: String)
 }