+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+* Do not fail when worldtime returns a fraction
+
 ## [2.1.1] - 2025-01-26
 
 ### Fixed